Do not allow empty ontology elements
This change improves the validation errors produced by RelaxNG
schema validators when an invalid event type, object type, source
or concept definition is encountered. The children of the ontology
elements used to be wrapped in <optional>, which caused the
validator to simply skip and ignore them if there was anything
wrong with their definition.